+++ src/main/scala/com/twitter/scalding/RichPipe.scala
@@ -293,6 +293,13 @@ class RichPipe(val pipe : Pipe) extends java.io.Serializable with JoinAlgorithms
   }
 
   /**
+   * Force a materialization to disk in the flow.
+   * This is useful before crossWithTiny if you filter just before. Ideally scalding/cascading would
+   * see this (and may in future versions), but for now it is here to aid in hand-tuning jobs
+   */
+  lazy val forceToDisk: Pipe = new Checkpoint(pipe)
+
+  /**
    * Convenience method for integrating with existing cascading Functions
    */
   def each(fs : (Fields,Fields))(fn : Fields => Function[_]) = {