+++ src/main/scala/com/twitter/scalding/TupleConversions.scala
@@ -27,7 +27,7 @@ trait TupleConversions extends GeneratedConversions {
   def toKeyValueList(tupe : TupleEntry) : List[CTuple] = {
     val keys = tupe.getFields
     (0 until keys.size).map { idx =>
-      new CTuple(keys.get(idx), tupe.get(idx))
+      new CTuple(keys.get(idx), tupe.getObject(idx))
     }.toList
   }
 
@@ -41,6 +41,13 @@ trait TupleConversions extends GeneratedConversions {
     else {
       Stream.Empty
     }
+  }
+
+  def tupleAt(idx: Int)(tup: CTuple): CTuple = {
+    val obj = tup.getObject(idx)
+    val res = CTuple.size(1)
+    res.set(0, obj)
+    res
   }
 
   implicit object TupleEntryConverter extends TupleConverter[TupleEntry] {
@@ -58,7 +65,7 @@ trait TupleConversions extends GeneratedConversions {
     if(iterable == null) {
       None
     } else {
-      scala.collection.JavaConversions.asScalaIterable(iterable)
+      scala.collection.JavaConversions.iterableAsScalaIterable(iterable)
     }
   }